feat: add authorization token for expiration

+++ src/pages/api/[...paths].ts
@@ -13,6 +13,12 @@ app.route('/secrets', secretController);
 app.get('/health', c => c.json({ success: true }, 200));
 
 app.get('expire', async c => {
+  const authorization = c.req.header('authorization');
+
+  if (authorization !== `Bearer ${import.meta.env.EXPIRATION_SECRET}`) {
+    return c.json({ success: false }, 403);
+  }
+
   await urlRepository.deleteAllExpired();
 
   return c.json({ success: true }, 200);
